backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
groq==0.15.0
mysql-connector-python==9.1.0
faiss-cpu==1.9.0.post1
numpy==2.1.3
sqlparse==0.5.2
